Splits frontend settings to individual components
