+++ frontend/app/src/components/settings/frontend/ScrambleDataSetting.vue
@@ -1,0 +1,29 @@
+<template>
+  <settings-option
+    #default="{ error, success, update }"
+    setting="scrambleData"
+    session-setting
+    :error-message="$tc('frontend_settings.validation.scramble.error')"
+  >
+    <v-switch
+      v-model="scrambleData"
+      class="general-settings__fields__scramble-data"
+      :label="$t('frontend_settings.label.scramble')"
+      :success-messages="success"
+      :error-messages="error"
+      @change="update"
+    />
+  </settings-option>
+</template>
+
+<script setup lang="ts">
+import { onMounted, ref } from '@vue/composition-api';
+import { set } from '@vueuse/core';
+import { getSessionState } from '@/composables/session';
+
+const scrambleData = ref<boolean>(false);
+onMounted(() => {
+  const state = getSessionState();
+  set(scrambleData, state.scrambleData);
+});
+</script>
